usb: misc: xapea00x: perform platform initialization of TPM
Normally the system platform (i.e., BIOS/UEFI for x86) is responsible
for performing intialization of the TPM.  For these modules, the host
kernel is the platform, so we perform the initialization in the driver
before registering the TPM with the kernel TPM subsystem.

The initialization consists of issuing the TPM startup command,
running the TPM self-test, and setting the TPM platform hierarchy
authorization to a random, unsaved value so that it can never be used
after the driver has loaded.